requirements.txt
num2words==0.5.10
praw==7.1.0
pytz==2020.1
requests==2.24.0
python-dotenv==0.14.0
PyYAML==5.3.1
